Remove `xla::ifrt::DeserializeOptions`
The way `xla::ifrt::DeserializeOptions` is currently implemented has a flaw, where the caller of `Deserialize()` must know the expected type to be deserialized. This defeats the purpose of the type erasure design in `xla::ifrt::Serializable` and makes it less useful.

`xla::ifrt::XlaProgram` was the only user of this feature and it was used to pass `mlir::MLIRContext`. Instead, `XlaProgram` deserialized from `Deserialize()` now also owns `mlir::MLIRContext` in addition to the module so that the caller of `Deserialize()` does not need to pass any options.

PiperOrigin-RevId: 543969105