Make some checks in FusionCanShareBufferHint explicit.
Before we relied on the order in which the HLO computation was traversed and
not checking a HLO op again whether it is elementwise. For the special case of
a DynamicUpdateSlice root that accesses the same slice as a DynamicSlice, we
could be sure that we will visit the root first coming from the first operand.
With the new code, we do not rely on this and make the checks more explicit.
Also allow it if the slices that are accessed are effectively a scalar.

PiperOrigin-RevId: 543983734