Change `xla::ifrt::Compiler::Compile` to take `std::unique_ptr<Program>` instead of `const Program&`
Taking `Program` as a const ref has some implications since IFRT is not strict about the use of `const`. For example, if we are to serialize the program inside `Compile`, we cannot do that today since `Serialize()` takes a reference (not const). Using `unique_ptr` aligns with the rest of IFRT API's look and feel.

PiperOrigin-RevId: 544383668