Fix --config=dbg in openxla/xla repo.
"-c dbg" works for the xla repo, but not the tensorflow repo. Currently, --config=dbg works for tensorflow repo, but not the xla repo (--config=dbg is still defined in xla, but it doesn't work). It's hard to remember to use the right flag in the right repo. This PR makes --config=dbg work in the xla repo so that can be used in both tensorflow and xla.

Fixes https://github.com/openxla/xla/issues/3381.

PiperOrigin-RevId: 544435135