Add an experimental ShapeAssertionOp.
The ShapeAssertionOp can be used to check that shape
dimension sizes meet a constraint. The constraint
depends only on dimension sizes, and can be evaluated
to a constant once all the shapes in the module are known.

The StablehloCanonicalizeDynamismPass removes these
operations if the constraint is met, and raises an
error otherwise.

PiperOrigin-RevId: 545015214