Directly create bitcasts instead of reshapes if we know they are bitcasts.
The ReshapeDecomposer pass runs after layout assignment. So when we decompose a
reshape to make it a bitcast, we can directly create a bitcast instead of
generating a reshape that might or might not be replaced by a bitcast
(depending on if we run another time the AlgebraicSimplifier pass).
As we are running the AlgebraicSimplifier again if we do layout normalization,
this change only affects cases where Layout normalization is disabled.

We usually assume in several places in our code base that a Reshape is not elementwise, but bitcasts are considered elementwise in some locations. We would either have to update all those locations to also check for Reshape bitcasts, or we need to make sure that we always replace reshapes with bitcasts if possible.

PiperOrigin-RevId: 545351553