Remove the CopyFusion pass, it is still buggy.
One bug is that we should not fuse a copy into a fusion if it can potentially
be affected by race condition. Also, we should not fuse a copy into a
multi-output reduction fusion where it is copying a reduce op.
These bugs can be fixed, but it takes some time, so until it is fixed, remove
the pass.

PiperOrigin-RevId: 545592296