[XLA:GPU] Outline cuBLAS padding requirements; make cuBLAS/Triton GEMM decision more robust.
The present order of passes in the compiler (padding for cuBLAS -> ... -> Triton GEMM rewrite -> cuBLAS GEMM rewrite) has a problem that padding for cuBLAS skips padding for those GEMMs which will be executed with Triton GEMM based on a complex check; if the condition for this check changes after that and before Triton GEMM rewrite and Triton rewriter skips that GEMM we end up with one that has to be executed by cuBLAS but is not padded. This change solves the problem by catching all non-padded ones in the Triton GEMM rewrite.

PiperOrigin-RevId: 545729159