[IFRT] Split OpaqueSharding into ConcreteSharding and ConcreteEvenSharding
IFRT `OpaqueSharding` previously handled two cases: (1) fully opaque (no
disassembly) (2) partially opaque (disassembly using a custom disassembly
function). The second case is slow to construct (the disassembly function
expects the shape of all shards), and is hard to serialize. Today, we only
expect the same shape for all shards, making it not meaningful to collect all
shard shapes.

This CL introduces two more specific shardings:

* `ConcreteSharding`: It is a type of opaque sharding that it supports
disassembly for a certain shape and not shard mapping to the global shape. This
is a drop-in replacement of all current use cases of `OpaqueSharding`. This
does not solve the slow construction and serialization of `OpaqueSharding`,
while it adds a more structure to the semantics with an extra check on the
assembled shape.

* `ConcreteEvenSharding`: This is a specialized and optimized version of
`ConcreteSharding` that uses the same shard shape for all shards. Most current
use cases can use `ConcreteEvenSharding`, with a few exceptions such as XLA
dynamic shape use cases.

`OpaqueSharding`'s role is now reduced to expresing a completely opaque
sharding that cannot be disassembled at all.

Most `OpaqueSharding` is migrated into `ConcreteSharding`. They will be
migrated to use `ConcreteEvenSharding` (or `HloSharding` once it can be used)
whenever safety can be ensured by construction or verified at runtime.

PiperOrigin-RevId: 546069138