[XLA:GPU] Require sm80 for cudnn_fused_conv_rewriter_test.
This test skips some testcases when running on pre-Ampere GPUs.  To ensure that
it's not flaky, we need to mark it as requiring Ampere.

We also change the test so it doesn't skip tests it doesn't need to.  In
particular, if you're just *compiling* but not *running* the operation, then
you shouldn't care what GPU is in the machine.

PiperOrigin-RevId: 546913352