PR #61230: Fix ambiguity in use of overloaded functions in XLA
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/61230

gcc complains of abiguity in some overloaded functions so cast the parameter to overcome this.
Copybara import of the project:

--
10785e9ca342539148464da2b959c7e61b319d62 by Andrew Goodbody <andrew.goodbody@linaro.org>:

Fix ambiguity in use of overloaded functions in XLA

gcc complains of abiguity in some overloaded functions
so cast the parameter to overcome this.

Merging this change closes #61230

PiperOrigin-RevId: 547107019