Add support for shape assertions to RefinePolymorphicShapes
Front-ends can use stablehlo.custom_call @shape_assertion to check that shape
dimension sizes meet a constraint that depends only on dimension sizes, and can be evaluated
to a constant once all the shapes in the module are known.

This is needed for jax2tf, because the JAX lowering relies
on constraints such as the batch dimensions of two inputs
being the same. These constraints are lost when lowering
to Stablehlo dynamic shapes, and it would be unsound to
invoke the code on inputs that do not meet the constraints.

We also add support for disabling this safety check if
"shape_assertions" is included in the `disabled_checks`.

PiperOrigin-RevId: 547482522