[XLA] Fix a bug in required assignment matching that caused buffers illegally being put in the alternate mem when inefficient allocations are enabled.
The combination of redundant eviction elimination and inefficient allocation
features caused a corner case bug. We had an overly conservative logic to find
previous allocations in AllocateSegment. When inefefficient allocation detection
inserted a required allocation in default memory, this logic sometimes failed to
find the correct previous allocation in the default memory. This CL makes this
logic less conservative.

This CL also adds a mechanism for tests to inject custom logic into inefficient
allocation detection in order to test complex scenarios like in this bug.

PiperOrigin-RevId: 548604568