[xla][gpu] Add AsyncStreamKind enum for GPU runtime streams.
Currently, the GPU runtime uses one stream to implement asynchronous collective
operations as well as Send and Recv P2P operations. We are planning to extend
this to two streams, one for asynchronous collective operations and one for P2P
operations.

To prepare for such a change, we first introduce the AsyncStreamKind enum and
use it to look up the corresponding runtime stream. We also extend the GPU
runtime to create one stream for each enum value.

PiperOrigin-RevId: 549376111