[xla][gpu] Use different resources to model asynchronous collective and P2P operations scheduling.
Previously, we use the same resource to model the scheduling of asynchronous
collective operations and P2P Send and Recv operations. This assumes we use
one runtime stream to implement these asynchronous operations.

We have just modified the runtime to user two streams, one for collective
operations and one for P2P Send and Recv operations. As such, we now modify the
HLO scheduler to use different resources to schedule collective operations and
P2P operations, to match the runtime implementation.

Add a test case.

PiperOrigin-RevId: 551036618