[TF:PJRT] Relax the required byte size check when the input shape is dynamic.
When the input shape is dynamic, returns an error if the require size of input buffer is smaller than the required size of expected execution shape.

An alternative solution is to always get the logical_on_device_shape but that calling logical_on_device_shape is blocking and has some performance impact.

PiperOrigin-RevId: 552508453