[XLA:GPU] Clean up cudnn runtime fusion, but don't enable it.
In a previous CL I disabled cudnn runtime fusion because we found some convs
that weren't working -- cudnn returned 0 algorithms, so you couldn't run the
conv.  It was unclear at the time why this was happening.

After reading the code today, I noticed that kElu convs have a check that was
missing in kRelu6 and kLeakyRelu convs.  Namely, kElu fusion skips convs with
an odd number of input/output features.

When I look at the failing testcases we had for kRelu6 and kLeakyRelu, they
both had an odd number of features.  Bingo!

Unfortunately, I cannot enable this by default, because it's very slow to
compile these convs on Ampere -- 7s with runtime fusion, as opposed to ~50ms
without.  This causes a 100x slowdown in compilation time.

PiperOrigin-RevId: 552642750