[IFRT] Make the internal state of `DeviceList` sharable
This change changes `DeviceList` to keep a shared internal state. This will
reduce expensive copies of `DeviceList` and enable the possibility of caching
for `DeviceList` that is useful if `DeviceList` would contain many `Device`s.

No change to the tests because this change does not affect the `DeviceList`
API. We will revisit if `std::shared_ptr<DeviceList>` is a more desirable form;
it will be easy to switch to that form.

PiperOrigin-RevId: 553605085