Registration mechanism for idempotent proto serializer.
The default proto serialization is not idempotent. This means
that two consecutive serializations of the same proto may not
be identical. Idempotent serialization is a requirement for
cache key generation.

Provide a mechanism that enables idempotent serialization.
If an idempotent serializer is available, it can be registered
and then used for serialization. If such a serializer is not
available, the non-idempotent serializer can be used as a
backup.

Also plumb the new serializer for the Python binding
CompileOptions.SerializeAsString.

Testing: new unit test, test workload.
PiperOrigin-RevId: 553918893