[XLA:GPU] Triton GEMM: decouple properties of block pointers.
Derive block pointer parameters separately for each input tensor so that one scope of a dot can access tensors with different shapes and layouts. This is ~NFC for now, it is required to enable more fusions but they will only come with other updates to the rewriter and the analysis.

PiperOrigin-RevId: 555414634