PR #61619: Remove dependency on simple_console_for_windows
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/61619

This is one way of getting around the problem where TF is too big to build on Windows. If bazel is told to create the runfiles symlink tree explicitly, Windows can use the same method as Linux to build the pip package without needing to zip everything up, and simple_console_for_windows is no longer needed.

The caveats are:

- The MSYS2 environment may need to be configured specifically to support symlinks; I am not sure. I set my test environment to use `MSYS=winsymlinks:nativestrict` but didn't test this solution without that.
- The build environment needs `rsync`, which replaces `cp` in one case where `cp` was giving me symlink-related errors.

I extracted this out of a work-in-progress environment I've been exploring, so I'm not 100% sure it will work right... but it seems very promising.
Copybara import of the project:

--
93596226a7df33dd085712d4bed9f66ee1b55e27 by Austin Anderson <angerson@google.com>:

Remove dependency on simple_console_for_windows

Merging this change closes #61619

PiperOrigin-RevId: 558183664