Avoid use after move in LayoutNormalization.
We can normalize the constant without moving the literal and creating a new
constant. This avoids the use after move in ReplaceInstruction() VLOG call.
Now that we do not create a new constant anymore, we need to make sure that we do
not revisit any of the normalized users of the constant. This should also be a
small efficiency boost, as now we do not re-normalize already normalized instructions.

Thanks to github user pemeliya for finding this bug.

PiperOrigin-RevId: 559385303