Fix NCCL UB issue
This is fixing a UB issue which occurs with newer version of Clang (16+).
The fix is also upstreamed through https://github.com/NVIDIA/nccl/pull/916.

In addition I'm changing a cc_library target to a cuda_library target.
This was an oversight from a previous change. I'm also correcting the version
number that we set in the patch - not sure if this version is reported in some
logs, but if it is, it should be correc.t

PiperOrigin-RevId: 561466895