[XLA:GPU] Speed up running light outside compilation ops.
Creating an OpKernel on each call is very slow.  So we add a hack to cache
these.  In fact, we leak them permanently!

The idea is to put a 128-bit hash of the TfCallbackData into the `opaque` of the
custom-call.  We use this to look up the OpKernel, in a hashtable.

An additional complexity is, we don't actually point to an OpKernel; rather, we
point to a structure which can contain multiple OpKernels.  When we run an op,
we "check out" an op kernel from a list, and return it when we're done.  This
way we can run the op concurrently (say, we're running the same model on
multiple GPUs at the same time).

This CL also caches the TensorShape objects, which we were previously creating
from the TfCallbackData.  This operation was surprisingly expensive; it didn't
look that huge in the profile, but once I got rid of it, a bunch of TCMalloc
calls disappeared, yay.

PiperOrigin-RevId: 562816602