[XLA:GPU][NFC] Add CreateMlirModuleOp as a leak-safe alternative of mlir::ModuleOp::create
mlir::ModuleOp::create returns a non-owning reference which is almost never the intended usage. We may leak memory if we don't assign it manually to an mlir::OwningOpRef. We actually had an error like this a few weeks ago.

CreateMlirModuleOp returns an owning reference by default.

I added a check to our internal presubmit, which will fail for mlir::ModuleOp::create calls.

We can opt-out of the check by adding /*ALLOW_MLIR_MODULE_OP_CREATE*/ to the same line as mlir::ModuleOp::create. I recommend only doing this if really needed and doing this in an utility function not in general code.

PiperOrigin-RevId: 566231363