[xla] Fix the HLO parser and verifier for wrapped kAsyncUpdate/Done instructions.
Previously, the HLO parser would created different wrapped async computations
for a pair of kAsyncStart/Update/Done instructions and rely on the HLO
verifier to make sure that the content of the computations are identical.
On the other hand, when a pair of async instructions have different async
computations, the pair of async instructions would be recorded at two async
computations' async_instructions_ separately. This crashes HloLiveRange
analysis as it expects both async instructions are recorded in the async
computations' async_instructions_.

We now change the parser to ensure that a pair of wrapped async instructions use
the same async computations. For unwrapped kAsyncStart/kAsyncDone
instructions, we keep the existing behavior and allow the instructions pair
to have different async computations in the parser and but report an error at
HLO validation.

Move a few tests from the verifier to the parser, as the parser now can't
accept such HLO inputs.

PiperOrigin-RevId: 567437233