[XLA] Fix big memory allocation and compile time slowdown for WhileLoopInvariantCodeMotion.
The pass allocates in some cases the input for the while-cond computation, but it really only
needs one of the tuple elements of the parameter, not the whole tuple.

Allocating the whole tuple (because the inputs to the loop can be big) can result in very large
memory allocations an significant time spent on memory allocation/initialization.

PiperOrigin-RevId: 567543997