Enable HLO with entry computation having token inputs and outputs.
Remove a check in the hlo_verifier that seems to be unnecessary anymore. This
enables JAX to emit HLO modules with tokens as inputs and outputs. Previously,
JAX would use tokens only for internal functions but for the module entry computation it would use bool[0], which would then be ignored and new tokens created for use with the inner functions.

This is needed in preparation for adding support for effects to JAX exporting, when we want to call a previously exported StableHLO module from another JAX function. For this, we need the tokens to use the same type both for inner functions and for the module main function.

PiperOrigin-RevId: 568257591