Allow quantized tensor as the input for mhlo.reshape. This also include change the folding logic to allow constant folding on quantized tensor.
There are basically two changes proposed here. The first change in reshape method is a fix to avoid folders to crash on quantized type, so that translation like MHLO -> (X) can run smoothly for quantized types. The second change in `materializeConstant allows constant folding for quantized types.

We note that the fixes are not ideal and there is a [ticket](https://github.com/openxla/stablehlo/issues/1691) to explore better solutions for the same.

PiperOrigin-RevId: 570133154