Fix TF_NCCL_VERSION detection.
The current macro substitution wasn't working and is resolving to %{nccl_version} rather than a version string. Instead, add code to create a NCCL config header (nccl_config.h), and use that to detect the .so version we should be opening.

This change should only affect NCCL when obtained via a stub, which is currently only used by an unreleased version of JAX.

PiperOrigin-RevId: 571081917