Fix determinism bug when Triton is used.
The issue was that the autotuner would choose between Triton and cuBLAS depending on which was faster. This behavior is now disabled when determinism is enabled.

Also add XLA determinism tests, including cases where Triton is not used. Before, determinism was only tested through TensorFlow tests but these tests do not reset the autotuning cache between runs of an HLO module and so can not catch autotuning bugs.

I didn't use HloTestBase's RunMultipleTimes method because it doesn't reset the GPU autotuning cache, and cannot reset it because I don't think we can add a dependency from HloTestBase to GPU-specific code.

PiperOrigin-RevId: 572980526