Capture cuBLAS logs for every API call
This installs a cuBLAS logger callback to capture cuBLAS logs for the duration
of all calls to the cuBLAS API that happen through CUDABlas::DoBlasInternalImpl.
If an error occurs, the captured logs are included as part of the error message
in an attempt to provide more context to the user.