Add CPU/GPU support for int4 conversions in XLA.
int4 is currently only supported for convert instructions and a few instructions that do not produce any GPU code, such as tuple.

int4 is not yet supported in the Triton emitter, so using int4 will currently be slow on GPUs (and it is likely slow on CPUs as well).

int4 is packed on device but not on host. To represent int4 being packed, a pass, SubByteNormalization, sets element_size_in_bits() on every int4 layout to 4. Care is taken to ensure subsequent passes do not create int4 layouts with element_size_in_bits unset.

PiperOrigin-RevId: 575041571