[xla:gpu] Remove silent recovery from failed CUDA graph updates
Trying to silently recover from CUDA graph update errors does not work well in practice as it can silently replace an error with a performance degradation, or even worse, another error.

Instead we should fix the root cause of these errors and honestly report the status up the call chain.

PiperOrigin-RevId: 575344185