Do not use the same computation for reduce and all-reduce
All-reduce computations are treated differently since #6126.
When a reduce op and an all-reduce op use computations that are functionally the same, the HloComputationDeduplicator pass will leave just one, as they compare as equal. (they appear the same in the HLO view)

Note: modifying HloComputation::Equal may have side effects, so only the compiler pass is updated.
PiperOrigin-RevId: 577953901