[XLA:GPU] Don't filter kernels which spill registers if we have only one kernel
When I ran our internal test with xla_gpu_triton_gemm_any enabled, I found some fusions which only try one tiling, but that tiling spills registers (4 bytes), so it is discarded and there is no kernel in the end, so compilation fails. To avoid this, we don't filter kernels which spill registers if we have only one kernel.

We should probably improve this in a more advanced way, but I think that this should be fine for now, to let us enable xla_gpu_triton_gemm_any soon.

PiperOrigin-RevId: 578771779