Do not set CUBLASLT_MATMUL_DESC_FAST_ACCUM, for now.
cuBLAS LT has a flag, CUBLASLT_MATMUL_DESC_FAST_ACCUM, that can be set for FP8 gemms. This flag causes the matmul to run faster but with lower accumulation precision. NVIDIA recommends using this flag on the forward pass on FP8 models but not the backwards pass, since the backwards pass needs more accumulation precision.

The PR https://github.com/openxla/xla/pull/6599 enabled fast accumulation on FP8 dots whose PrecisionConfig is DEFAULT (but not HIGH or HIGHEST). This allows layers in frameworks to use fast accumulation on the forward pass but not the backwards pass by setting the PrecisionConfig on the backwards pass to be HIGH or HIGHEST.

The issue is, Flax and Praxis do not yet set the PrecisionConfig to HIGH or HIGHEST on the backwards pass, so the PR will cause poor FP8 training quality. The PR should not have been merged until Flax and Praxis set the PrecisionConfig, but I didn't realize this and merged it anyway. Reverting the PR is a pain, so instead this CL just removes the line that sets CUBLASLT_MATMUL_DESC_FAST_ACCUM, while keeping most of the plumbing around it. This CL will be rolled back once Flax and Praxis set the PrecisionConfig.

PiperOrigin-RevId: 579018421