Support int4 arrays with an odd number of elements.
Before, an odd number of elements was not supported since the loop in parallel_loop_emitter.cc had an unroll_factor of at least 2 for int4 arrays. Having the loop unrolled meant we might compute and write an element with an out-of-bounds index. To fix this, a bounds check is added to each iteration of the unrolled loop. This adds overhead and prevents any possibility of vectorization, but is better than not supporting int4 arrays with odd numbers of elements.

PiperOrigin-RevId: 579282947