[XLA:GPU] Add xla_gpu_force_autotuning_parallelism flag
This helps to test how long would be the compilation on a machine with fewer CPU threads.

An earlier CL modified the logic so that xla_gpu_force_compilation_parallelism doesn't affect the autotuning parallelism anymore:
https://github.com/openxla/xla/commit/61a7666906f384bf08af1eb0eef9d8ace88be484

Because of this, we had no way to specify autotuning parallelism, without this CL.

It would be problematic to simply modify xla_gpu_force_compilation_parallelism to affect autotuning as well, because some users set that to 1 and autotuning would become very slow for them.

So now if we really don't want any multi-threading, we should set:
--xla_gpu_force_compilation_parallelism=0
--xla_gpu_force_autotuning_parallelism=0
PiperOrigin-RevId: 580186689