Introduce stablehlo/experimental
This patch cleanly separates the experimental extensions to the StableHLO repository added in openxla/xla and tensorflow/mlir-hlo.

Previous implementation of this notion had the non-experimental and experimental pieces mixed together which complicated ongoing development and downstream integrations.

In this patch, we only copy/paste the implementations of --stablehlo-canonicalize-dynamism and --stablehlo-refine-shapes into stablehlo/experimental/transform to simplify review. This duplication will be removed in a follow-up patch.

PiperOrigin-RevId: 580640801