[PJRT] Remove unused EnumerateDevices API.
The users of EnumerateDevice() were refactored to use the KV-store instead as part of the PJRT plugin work. There are no remaining users.

Also rename "enumerate_devices_timeout" to "cluster_register_timeout" which matches the name given in the coordination service code. Bump the default timeout to one hour, which also matches the coordination service default.

PiperOrigin-RevId: 582058387