[PJRT] Allow the compiler to choose executable input and output layouts.
This is done by adding a new "mhlo.layout_mode" attribute to arguments
and results in the input MLIR module to the compiler. The layout can
either be the default (i.e. compact shape), compiler chooses, or caller
chooses by specifying a layout.

Prior to this change, the caller could specify layouts via
CompileOptions.argument_layouts and
CompileOptions.executable_build_options.result_layout, and otherwise
the default compact layout would be used. This change both introduces
the third compiler-chooses option, and embeds this info in the IR
instead of as a separate compile option. The old options still work
while we update callers and take precedence if specified.

PiperOrigin-RevId: 582082768