[xla:gpu] Add gpu-schedule-postprocessing pass.
Add a boolean field, no_parallel_gpu_op, to CollectiveBackendConfig. This field
asserts that an asynchronous collective operation does not execute in parallel
with other operations in GPU. The default value of the attribute is false,
which should lead to conservative runtime behavior.

Add BackendConfig test for the field.

Add gpu-schedule-postprocessing pass, to refine the attribute value. Add test
cases for the pass.

PiperOrigin-RevId: 582457930