[XLA] Match other dimensions to try to swap instead of just directly matching replicated dimensions when the current replicated dimensions doesn't match.
It's not necessarily faster to match a non-matching replicated dimension when grouping NonContracting dimensions
and it can cause significant reshuffling of data. Instead try to match another dimension that if it was the replicated one
then we would be able to match. If we find it the reshuffling is much more regular (and over specific axes) than the random one introduced by just blindly matching the replicated dimension.

PiperOrigin-RevId: 582838552