[XLA:GPU] Don't generate unreachable operations in fusions in edge cases when some instructions are reachable from both operands, and they are not fusible from the first operand, but fusible from the second.
I tried to enable Triton fusions for some new cases, and I found that some of the fusions contained dead code (unreachable instructions).
The reason was that some instructions were reachable both through the LHS and the RHS.
The BFS (Breadth-first search) through the LHS operand "marked" one operation as non-fusible because it would exceed the limit on fusion parameters per operand.
But the BFS through the RHS operand went through that node and fused some more operands.
So the resulting fusion was not connected and caused errors.
This CL fixes the issue by not visiting again instructions which were decided to not be fused in previous scopes.
Sorry for the complicated test case.

Extended TritonGemmTest.SameInput to make sure that this solution doesn't prevent using the same fused instructions for LHS and RHS.

PiperOrigin-RevId: 583319776