[xla:gpu] Revise gpu-schedule-postprocessing pass to exclude fusion.
This is because cuLaunchKernel is safe, even though it could technically block
if too many unsubmitted commands are in the push buffer but this doesn't happen
in practice. Also rename no_parallel_gpu_op to no_parallel_custom_call.

Modify the added routines to use a const parameter when it is appropriate.

This is to address the follow up comments for the PR that adds the pass.

PiperOrigin-RevId: 583510144