[XLA] Avoid making an extra copy of constant literals when compiling to llvm IR.
Before this change the constants were first copied from HLO to LHLO MLIR, and then from LHLO MLIR into an intermediate buffer - `GpuExecutable::ConstantInfo`. This change turns the intermediate buffer into an alias to the data stored in the HLO graph in most cases. In the case of 4-bit integers we still need to make a copy since the needed data format is different. However, 4-bit ints are very rare.

Especially with large constants, this change significantly reduces the peak memory usage.

The copy of the data in LHLO MLIR remains, even tough it has a shorter lifetime than the HLO data. In theory, it could be possible to avoid also the copy in the LHLO MLIR, but that I have not looked into how much effort that could be (likely a lot).

PiperOrigin-RevId: 584107265