Fix an incorrect static_assert
The size of a uint8_t is 1 so a static_assert
to check that it is 0 will always be false.
The original intent was to have the assert only
trigger if the struct was instantiated but the
standard deems it ill formed if it can never be
true and allows compilers to reject it. Adopt a
different workaround that avoids this by allowing
the possibility of an evaluation to true.
Also fix a couple of warnings about lack of
typename