Replace boundary functions with a fusion adaptor class.
Boundary functions seemed like a nice and easy abstraction for fusions,
but they turned out to be too difficult to use in practice. The main
problem is that everything is still based on HloInstructions, whose
users and operands are difficult to traverse in general.

The solution introduced here is to introduce an HloFusionAdaptor class
with a simple interface, and an HloInstructionAdaptor which always behaves
as if the HLO was completely unfused.

If I had more time, I would have made smaller change.

PiperOrigin-RevId: 585087631