[XLA:GPU][NFC] Transform dimension propagation to the functional paradigm
This intends to make it easier to implement new fusion strategies, such as "Separate multiple uses of nodes within one scope when they are incompatible in Triton GEMM fusion".

"Renames":
AnalyzeForFusion -> GetPropagatedDimOrdersAndRequirementsIfProfitablyFusible
RequireSupportedInstruction -> GetPropagatedDimOrdersAndRequirements
HandleInstruction -> GetPropagatedDimOrders
RequireSupportedDimOrder -> GetRequirementsIfSupportedOrder
RequireSupportedDimOrders -> GetRequirementsIfSupportedOrders
DimOrderUpdates -> DimOrdersAndReqs

Notable logic changes:

I split out the splittable_dimension_major_part_size from DotProperties to DotRequirements, because it's not really a property of the dot, but rather a requirement which can be imposed by the instructions of the fusion.

I explicitly return an error if a dimension split would be needed for Softmax in GetRequirementsIfSupportedOrder.

I don't check IsSupportedSplittableDimensionMajorPartSize in GetRequirementsIfSupportedOrder anymore, I just check that in CombineDimOrdersAndReqs after the propagation is done.

PiperOrigin-RevId: 585650283