[XLA][NFC] Mark all singular DebugOptions fields as explicitly optional.
proto3 introduced implicit presence, in which the has-bit is not stored for
singular scalar fields and thus it is impossible to tell the difference between
an unset field and a field explicitly set to its default value. This makes it
very hard to merge two DebugOptions instances, especially when one comes from
flags where the flag-level default may differ from the proto-level default.

A unit test enforces this to prevent regressions.

PiperOrigin-RevId: 585787781