Lower shape.cstr_broadcastable op in ShapeLegalizeToHLO
Lower the op to shape_assertion custom_calls. And replace the result shape.witness with a const_witness of true. This allows the subsequent shape.assuming regions to be removed.

Also add an option to the pass to control whether to lower cstr_xxx ops. This makes the impl easier as we can reuse util functions in the file. And it  makes sure existing use cases are not modified.

PiperOrigin-RevId: 587167546