[XLA:GPU] Separate multiple uses of nodes in Triton GEMM fusions
This is a "logic cleanup" plus it allows a few more fusions when some HLOs must be iterated in different ways when we use them multiple times in the fusion.

Later we can make it so that they are only separated if the DimOrders are different for the different uses.

The main function "CreateDotFusion" is now much cleaner.

PiperOrigin-RevId: 588108179