[XLA:GPU] Emit llvm.minimum/maximum if we have sm_80
This propagates NaNs according to our desired semantics (if any input is NaN,
the output is NaN) and is quite a bit faster. Sadly it canonicalizes NaNs to a
value that no other platforms uses, so put an extra canonicalization step into
the test.

As far as I can tell cost model already models this as "single instruction"
which was actually wrong for the default lowering.

PiperOrigin-RevId: 588344430