[XLA:GPU] Refactor fusion_merger_triton to prepare for consumer merging pass
Currently the producer fusion merging pass is called by the main FusionMergerTriton::Run method, and returns a bool representing success. However, in order to add a consumer fusion merging pass, we'd rather return a StatusOr<HloInstruction> with a pointer to the merged function, so we can use that pointer when merging in the consumer fusion.

This CL creates a new method which calls the the producer merging pass and handles the StatusOr<HloInstruction>, so that the consumer fusion merging pass can be added with minimal refactoring in that CL.

PiperOrigin-RevId: 588441942