[XLA:GPU] Fix crash in triton tiling of bitcast ops.
Bitcast ops may try to split a dimension across batch and reduction dimensions in softmax. This is unsupported because the tiling fragments cannot be propagated to represent both batch and reduce dims in the same literal dimension, and it leads to a crash in TritonFusionAnalysis.
This change checks if the reduce dimension shape is not retained and gracefully fails in this case. It relies on the assumption that the softmax reduce dim will always be minormost.

Reverts 7245d6ce911587f7fbfdeab2a0c5d1bb3d724989

PiperOrigin-RevId: 588584412