[XLA:GPU] Fix crash in triton tiling of bitcast ops.
Bitcast ops may try to split a dimension across batch and reduction dimensions in softmax. This is unsupported because the tiling fragments cannot be propagated to represent both batch and reduce dims in the same literal dimension, and it leads to a crash in TritonFusionAnalysis.
This change fixes the crash by checking if the remainder of a dimension propagated from reduce dims also gets propagated into a batch dim fragment, and gracefully failing analysis in this case.

PiperOrigin-RevId: 588877911