[XLA:GPU] Move cache invalidate code into a helper.
Call the helper at PreFusion. The benefit is that at PreFusion, both producer and consumer are still valid instruction and we can delete caches from users/operands.

At OnFusingInstruction, original_producer and original_consumer can be destroyed and detached from users and operands, so it's harder to find cache entries that need to be removed.

PiperOrigin-RevId: 589846248