[XLA:GPU] Use f32->bfloat conversion instructions on sm_80+
We tried this before with an intrinsic, but that breaks vectorization. Relying
on native LLVM types doesn't while delivering the same code improvements. The
downside is that LLVM now knows that it's a bfloat instead of a i16 and will
optimize based on it. While making this change I had to patch a bunch of holes
in the NVPTX LLVM backend, there might be more.

Depends on https://github.com/llvm/llvm-project/pull/74827

PiperOrigin-RevId: 590118269