[xla:gpu] Add a specialized kTraced state to command buffers
A command buffer created by tracing can't be updated, so we create a kTraced state which can't transition to other states.

This CL fixes a memory leak:
We should not call GpuDriver::CreateGraph when constructing a command buffer for tracing. Because stream capture creates a new cuda graph, and the old empty graph becomes a dangling pointer.

PiperOrigin-RevId: 590172371